Update for new version of Skype4J and initial ground work for syncing previous messages

+++ src/main/java/pro/zackpollard/telegrambot/skypetotelegrambot/storage/CredentialStore.java
@@ -13,7 +13,7 @@ import java.util.Map;
 public class CredentialStore {
 
     @Getter
-    private final Map<Integer, Credentials> telegramToSkypeCredentials;
+    private final Map<Long, Credentials> telegramToSkypeCredentials;
 
     public CredentialStore() {
